In the beginning, Node said '200: Hello World!'
We start out with a simple http server in vanilla nodejs. At this
moment the only dependency is node. If you have installed node on your
system, this server can be started with the command "node app.js".

This starts a http server on http://localhost:8080/ which will respond
with http status code 200 (which means everything is OK) and a
Content-Type of 'text/plain'.

So please start the server and point your browser (or curl) to
http://localhost:8080 to verify your installation.