Installing Express using NPM
We would like to use the Express framework which will make developing
our web server a lot more convenient. It abstracts difficult things and
provides lots of advanced functionality out of the box with a clear API.

To install express, and future packages, we use the node package manager
(npm). The npm executable checks if there is a package.json file in your
project and if you use the command "npm install" it will attempt to
install all the dependencies (recursively) that you gave in your
package.json file.

Your application can also be made available to others. If you decide to
do so later, other people will be able to add "hello-git-express" to
their list of dependencies. However, we will set this package to private
for the time being.

You can install the dependencies now using "npm install".