Using Express in our code
So that concluded the setup part. Now that we are indeed set up
properly, we can start using Express.

We don't need the standard http server anymore, express will take care
of that. (It probably uses the http server inside, you can always just
check the source by going into your node_modules folder.)

Instead of requiring http, we now require express. The express server is
created in a different way, instead of exposing a createServer method,
the module itself can be called, which returns an instance of an express
server.

An express server starts listening the same way the standard http server
starts listening: we call listen(8080) on the object.

But how do we send responses? Well, we now limit request to /. If we try
to access /whatever, we will get an error.