Configuration done right
We would like to be able to configure things like compression,
internationalization, constants, port numbers, etc.

To do this, Express exposes the configure method, which in this form is
basically a direct callback. Why would we use this then? Well, it is
only ordering your configuration in one code block, which is not a
really good reason, but it is easier to read.

However, the configure method is not useless, it has another use which
allows you to configure development/testing/acceptance and production
environment in different ways. We shall see this later.

Right now we simply configure the port number.