Static resources
We need a way to serve static resources, like images. Express allows you
to do so by using middleware. The concept is a bit vague, I'll try to
explain.

In Express, each request passes through a list of middleware functions,
in predefined and strict order. Each middleware function gets the power
to do with this request what it wants:
 * it can process the request and return
 * it can modify the request and pass it on to the next middleware
 * it can ignore the request and pass it on to the next middleware

Right now, we want to register the built in "static file handler"
middleware. We register it, like all middleware functions, using the
use method.

We configure the static file handler in such a way that it tries to see
if the requested file exists on disk, under the directory 'static'. To
make this work, we create a directory static and place a file in there
called static.html .

If we now go to localhost:8080/ we still see 'Hello World'. If we go to
localhost:8080/static.html, we see 'Hello'.