Middleware order
The app.get('/', func..) statement is configuration for the Express
router. This is another piece of middleware.

If you now were wondering why the request to '/' still gave any results,
or why it ever gave any results at all, you score bonus points.

In a perfectly consistent world, we would have needed to register the
Express router middleware before something like the route registration
app.get('/', func..) would work.

It turns out that if the router was not registered, the router is
registered automagically when you register a route, like we did with
'/'.

So, in our previous example, the router was registered after we
registered the static middleware. This comes with a performance penalty,
as the static middleware (and thus, the disk) will now be hit for each
request. To fix this, we register the router explicitly before the
static middleware.