Dynamic views and templates
We need a bit more than static files and simple text responses. We can
solve this by using a view engine, which needs to be configured. We need
to set a directory that contains the views, we need to set the type of
view engine and we need to add the view engine to our dependencies in
package.json.

Because we added something to package.json, we need to run 'npm install'
once more.

Once everything has been configured, we can render these templates using
res.render. The function requires two parameters, one with the name of
the view, one parameters object. Each of the attributes in this object
are available in the view. Here we send a time attribute, which we use
in the index.jade view.

In express we can choose between multiple view engines. I like jade, so
I chose that.